+++ htl-services/src/main/java/com/healthline/services/api/ITimelineService.java
@@ -4,7 +4,6 @@
 package com.healthline.services.api;
 
 import java.io.InputStream;
-import java.math.BigInteger;
 
 import com.healthline.entity.Event;
 import com.healthline.entity.Timeline;
@@ -16,11 +15,11 @@ import com.healthline.entity.Timeline;
 public interface ITimelineService
 {
 
-    void createTimeline(BigInteger userId, String headline, String description);
+    void createTimeline(Long userId, String headline, String description);
 
-    void addEventToTimeLine(BigInteger timelineId, Event event, String fileName, InputStream fileData);
+    void addEventToTimeLine(Long timelineId, Event event, String fileName, InputStream fileData);
 
-    Timeline getTimeline(BigInteger userId);
+    Timeline getTimeline(Long userId);
 
     String deleteTimeline();
